#137 [FEAT] : 인기있는 장소 결제내역 횟수 체크 스케줄러 추가

+++ Backend/src/main/java/com/example/backend/domain/popular/dto/PopularResponseDto.java
@@ -12,7 +12,7 @@ public class PopularResponseDto {
     @Getter
     @Builder(toBuilder = true)
     public static class PopularPlace {
-       private Integer popularId;
+       private Long popularId;
        private String store;
        private String img;
        private String category;
